requirements.txt
torch
torchvision
transformers
pandas
tqdm
ftfy
regex
Pillow
numpy
wandb
PyYAML
git+https://github.com/openai/CLIP.git